+++ src/main/java/faang/school/godbless/Task_FanOut_FanIn/ResultConsumer.java
@@ -1,0 +1,17 @@
+package faang.school.godbless.Task_FanOut_FanIn;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+public class ResultConsumer{
+
+    private final AtomicLong sumOfSquaredNumbers;
+    public ResultConsumer(Long init) {
+        sumOfSquaredNumbers = new AtomicLong(init);
+    }
+    public Long add(final Long num) {
+        return sumOfSquaredNumbers.addAndGet(num);
+    }
+    public Long getValue() {
+        return sumOfSquaredNumbers.longValue();
+    }
+}
